Fix format string warning in DEBUG_HEX_DUMP macro
- Changed format specifier from %d to %zu for size_t values
- Added explicit cast to size_t for len parameter
- Eliminates compilation warnings when using sizeof() with DEBUG_HEX_DUMP
- Updated debug binaries with warning-free compilation

+++ lan8651_ethtool.c
@@ -33,7 +33,7 @@
 #define DEBUG_ENTER() DEBUG_PRINT("ENTER")
 #define DEBUG_EXIT(ret) DEBUG_PRINT("EXIT with %d", ret)
 #define DEBUG_HEX_DUMP(data, len) do { \
-    DEBUG_PRINT("Hex dump (%d bytes):", len); \
+    DEBUG_PRINT("Hex dump (%zu bytes):", (size_t)len); \
     for(int i = 0; i < len; i++) { \
         fprintf(stderr, "%02X ", ((unsigned char*)data)[i]); \
         if((i+1) % 16 == 0) fprintf(stderr, "\n"); \
